+++ src/TestableHttpClient/Utils/HttpRequestMessageCloner.cs
@@ -1,6 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-namespace TestableHttpClient.Utils;
+﻿namespace TestableHttpClient.Utils;
 
 internal static class HttpRequestMessageCloner
 {
@@ -18,12 +16,13 @@ internal static class HttpRequestMessageCloner
             clone.Headers.TryAddWithoutValidation(item.Key, item.Value);
         }
 
-        // Copy content (buffered)
         if (request.Content is not null)
         {
-            var bytes = await request.Content.ReadAsByteArrayAsync().ConfigureAwait(false);
+            var bytes = await request.Content
+                .ReadAsByteArrayAsync(cancellationToken)
+                .ConfigureAwait(false);
             var contentClone = new ByteArrayContent(bytes);
-
+            contentClone.Headers.Clear();
             // copy content headers
             foreach (var header in request.Content.Headers)
             {